requirements.txt
streamlit>=1.38.0
pandas>=2.2.2
numpy>=2.1.2
matplotlib>=3.9.2
seaborn>=0.13.2
openai>=1.43.0
scipy>=1.14.1
openpyxl>=3.1.5
xlrd>=2.0.1
google-generativeai>=0.8.5
